requirements.txt
Flask>=2.0
argon2-cffi>=21.0
pytest>=7.0
pytest-playwright>=0.5
playwright>=1.30
